refactor: 日付ユーティリティの重複を解消し api を型チェック対象に追加
