+++ src/lib/dateUtils.ts
@@ -1,17 +1,25 @@
 /**
  * Date utility functions for date parsing, validation, and formatting.
+ *
+ * 日付そのものの計算は timeService に委譲し、ここでは「ブラウザのタイムゾーンを
+ * 使う」というアプリ固有の判断と、表示用の整形だけを担う。
  */
+
+import {
+  getLocalDate,
+  getBrowserTimeZone,
+  addDays as addDaysToDate,
+} from '@/domain/services/timeService';
 
 const DATE_REGEX = /^\d{4}-\d{2}-\d{2}$/;
 
 const DAY_NAMES = ['日', '月', '火', '水', '木', '金', '土'] as const;
 
 /**
- * Get today's date as YYYY-MM-DD string.
+ * Get today's date as YYYY-MM-DD string, in the browser's timezone.
  */
 export function getTodayString(): string {
-  const now = new Date();
-  return formatToDateString(now);
+  return getLocalDate(new Date(), getBrowserTimeZone());
 }
 
 /**
@@ -63,9 +71,7 @@ export function formatDisplayDate(dateStr: string): string {
  * Add days to a date string and return a new YYYY-MM-DD string.
  */
 export function addDays(dateStr: string, days: number): string {
-  const date = new Date(`${dateStr}T00:00:00`);
-  date.setDate(date.getDate() + days);
-  return formatToDateString(date);
+  return addDaysToDate(dateStr, days);
 }
 
 /**
@@ -80,11 +86,4 @@ export function isToday(dateStr: string): boolean {
  */
 export function isFutureDate(dateStr: string): boolean {
   return dateStr > getTodayString();
-}
-
-function formatToDateString(date: Date): string {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-  return `${year}-${month}-${day}`;
 }
